Checklists & Automated PR for Migrations
This adds a few checklists to start to build out the process guidelines.
I set up the dangerfile to automatically add the database migrtion
checklist into any PR that includes it.

Additionally, there are some Rubocop tweaks to get it to line up with
the styles in use in the project.